+++ Purchases/Logging/Strings/ReceiptStrings.swift
@@ -17,25 +17,69 @@ import Foundation
 // swiftlint:disable identifier_name
 enum ReceiptStrings {
 
-    static let data_object_identifer_not_found_receipt = "The data object identifier couldn't be found " +
-        "on the receipt."
-    static let force_refreshing_receipt = "Force refreshing the receipt to get latest transactions " +
-        "from Apple."
-    static let loaded_receipt = "Loaded receipt from url %@"
-    static let no_sandbox_receipt_intro_eligibility = "App running on sandbox without a receipt file. " +
-        "Unable to determine into eligibility unless you've purchased before and there is a receipt available."
-    static let no_sandbox_receipt_restore = "App running in sandbox without a receipt file. Restoring " +
-        "transactions won't work until a purchase is made to generate a receipt. This should not happen in " +
-        "production unless user is logged out of Apple account."
-    static let parse_receipt_locally_error = "There was an error when trying to parse the receipt " +
-        "locally, details: %@"
-    static let parsing_receipt_failed = "%@-%@: Could not parse receipt, conservatively returning true"
-    static let parsing_receipt_success = "Receipt parsed successfully"
-    static let parsing_receipt = "Parsing receipt"
-    static let refreshing_empty_receipt = "Receipt empty, refreshing"
-    static let unable_to_load_receipt = "Unable to load receipt, ensure you are logged in to a valid " +
-        "Apple account."
-    static let unknown_backend_error = "Unexpected backend error when posting receipt. Make sure you " +
-        "are on latest SDK version and let us know if problem persists."
+    case data_object_identifer_not_found_receipt
+    case force_refreshing_receipt
+    case loaded_receipt(url: URL)
+    case no_sandbox_receipt_intro_eligibility
+    case no_sandbox_receipt_restore
+    case parse_receipt_locally_error(error: Error)
+    case parsing_receipt_failed(fileName: String, functionName: String)
+    case parsing_receipt_success
+    case parsing_receipt
+    case refreshing_empty_receipt
+    case unable_to_load_receipt
+    case unknown_backend_error
+
+}
+
+extension ReceiptStrings: CustomStringConvertible {
+
+    var description: String {
+        switch self {
+
+        case .data_object_identifer_not_found_receipt:
+            return "The data object identifier couldn't be found on the receipt."
+
+        case .force_refreshing_receipt:
+            return "Force refreshing the receipt to get latest transactions from Apple."
+
+        case .loaded_receipt(let url):
+            return "Loaded receipt from url \(url.absoluteString)"
+
+        case .no_sandbox_receipt_intro_eligibility:
+            return "App running on sandbox without a receipt file. " +
+            "Unable to determine into eligibility unless you've purchased " +
+            "before and there is a receipt available."
+
+        case .no_sandbox_receipt_restore:
+            return "App running in sandbox without a receipt file. Restoring " +
+            "transactions won't work until a purchase is made to generate a receipt. " +
+            "This should not happen in production unless user is logged out of Apple account."
+
+        case .parse_receipt_locally_error(let error):
+            return "There was an error when trying to parse the receipt " +
+           "locally, details: \(error.localizedDescription)"
+
+        case .parsing_receipt_failed(let fileName, let functionName):
+            return "\(fileName)-\(functionName): Could not parse receipt, conservatively returning true"
+
+        case .parsing_receipt_success:
+            return "Receipt parsed successfully"
+
+        case .parsing_receipt:
+            return "Parsing receipt"
+
+        case .refreshing_empty_receipt:
+            return "Receipt empty, refreshing"
+
+        case .unable_to_load_receipt:
+            return "Unable to load receipt, ensure you are logged in to a valid Apple account."
+
+        case .unknown_backend_error:
+            return "Unexpected backend error when posting receipt. Make sure you " +
+            "are on latest SDK version and let us know if problem persists."
+
+        }
+    }
 
 }