+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 4.0.0-beta.1
+- First beta for RevenueCat (previously Purchases) framework 🎉
+    100% Swift framework + ObjC support.
+- See our [RevenueCat V4 API update doc](docs/V4_API_Updates.md) for API updates.
+
 ## 3.12.3
 - Fixed a bug where checkTrialOrIntroductoryPriceEligibility would return `eligible` for products that don't have intro pricing
     https://github.com/RevenueCat/purchases-ios/pull/679